fix: proxy.tsにメンテナンスモードを統合（middleware.tsとの競合を解消）
- middleware.tsはproxy.tsと競合するため削除
- メンテナンスモードのロジックをproxy.tsに統合
- webapp/admin両方でproxy関数の先頭でメンテナンスモードをチェック
